Serialization is now done through a more abstract object, Packet
Packets can contain Vehicles, as well as contain information about what
type of request is being sent and search strings.

When the server receives a byte stream, it attempts to convert it to a
packet, upon failure it attempts a conversion to image file.

Icons are now in the table header

Reverted a change that made vehicle dating long types. They are now back
to date types for optimization.

new refresh button to be implemented soon